backendChallenge/pyproject.toml
[project]
name = "backendchallenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.2",
    "django-ninja>=1.4.3",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
]
